[#2827] Move metrics classes into service-base module.
