+++ adapter-base/src/main/java/org/eclipse/hono/adapter/resourcelimits/ConnectionDurationAsyncCacheLoader.java
@@ -20,7 +20,7 @@ import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
 
-import org.eclipse.hono.adapter.metric.MicrometerBasedMetrics;
+import org.eclipse.hono.service.metric.MicrometerBasedMetrics;
 import org.eclipse.hono.tracing.TracingHelper;
 import org.eclipse.hono.util.TenantConstants;
 import org.slf4j.Logger;
