+++ adapters/http-vertx-base/src/main/java/org/eclipse/hono/adapter/http/MicrometerBasedHttpAdapterMetrics.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018, 2020 Contributors to the Eclipse Foundation
+ * Copyright (c) 2018, 2021 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -13,7 +13,7 @@
 
 package org.eclipse.hono.adapter.http;
 
-import org.eclipse.hono.adapter.metric.MicrometerBasedMetrics;
+import org.eclipse.hono.service.metric.MicrometerBasedMetrics;
 
 import io.micrometer.core.instrument.MeterRegistry;
 import io.vertx.core.Vertx;
